requirements.txt
pyyaml==6.0.2
pydantic==2.9.2
